pyproject.toml
[project]
name = "langchain-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-google-genai>=4.2.1",
    "langchain-pinecone>=0.2.13",
    "pinecone>=6.0.0,<8.0.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
